+++ apps/web/src/components/recipe-list/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Recipe-list module barrel.
+ *
+ * Exposes the shared `RecipeListView` plus its leaf components, the
+ * URL-param hook, and the equipment-filter constants used by both
+ * `/recipes` and `/recipes/starred`.
+ */
+export { RecipeListView, type RecipeListViewProps } from './RecipeListView.tsx';
+export { FilterField } from './FilterField.tsx';
+export { ActiveFilterBadge } from './ActiveFilterBadge.tsx';
+export { RecipeCard } from './RecipeCard.tsx';
+export { PaginationControls } from './PaginationControls.tsx';
+export { useRecipeFilters } from './useRecipeFilters.ts';
+export { EQUIPMENT_FILTER_TYPES, EQUIPMENT_TYPE_LABELS } from './constants.ts';
+export type { EquipmentFilterType } from './constants.ts';
